+++ src/main/java/com/github/correa/finalreality/controller/handlers/IEventHandler.java
@@ -1,0 +1,6 @@
+package com.github.correa.finalreality.controller.handlers;
+
+import java.beans.PropertyChangeListener;
+
+public interface IEventHandler extends PropertyChangeListener {
+}