+++ src/main/java/com/github/correa/finalreality/enums/CharacterType.java
@@ -1,0 +1,8 @@
+package com.github.correa.finalreality.enums;
+
+/**
+ * Holds the types of characters.
+ */
+public enum CharacterType {
+  KNIGHT, ENGINEER, THIEF, BLACK_MAGE, WHITE_MAGE, ENEMY
+}