+++ src/main/java/com/github/correa/finalreality/model/character/player/classes/commonclasses/Thief.java
@@ -1,10 +1,13 @@
 package com.github.correa.finalreality.model.character.player.classes.commonclasses;
 
+import com.github.correa.finalreality.enums.CharacterType;
+import com.github.correa.finalreality.enums.Stats;
 import com.github.correa.finalreality.model.character.ICharacter;
 import com.github.correa.finalreality.model.character.player.AbstractPlayerCharacter;
 import com.github.correa.finalreality.model.weapon.IWeapon;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.HashMap;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
@@ -17,18 +20,18 @@ public class Thief extends AbstractPlayerCharacter {
    *  the queue with the characters waiting for their turn.
    * @param name
    *  the Thief's name.
-   * @param hitPoints
-   *  the Thief's hit points.
+   * @param maxHitPoints
+   *  the Thief's maximum hit points.
    * @param defensePoints
    *  the Thief's defense points.
    */
   public Thief(
       @NotNull BlockingQueue<ICharacter> turnsQueue,
-      @NotNull String name, int hitPoints,
+      @NotNull String name, int maxHitPoints,
       int defensePoints) {
     super(
         turnsQueue, name,
-        hitPoints, defensePoints);
+        maxHitPoints, defensePoints);
   }
 
   @Override
@@ -39,6 +42,15 @@ public class Thief extends AbstractPlayerCharacter {
   @Override
   public void equip(final @NotNull IWeapon weapon) {
     weapon.equippedByThief(this);
+  }
+
+  @Override
+  public HashMap<Stats, String> getInfo() {
+    super.getInfo();
+    info.put(
+        Stats.CHARACTER_TYPE,
+        String.valueOf(CharacterType.THIEF));
+    return info;
   }
 
   @Override