+++ src/main/java/com/github/correa/finalreality/model/character/player/classes/mageclasses/WhiteMage.java
@@ -1,10 +1,13 @@
 package com.github.correa.finalreality.model.character.player.classes.mageclasses;
 
+import com.github.correa.finalreality.enums.CharacterType;
+import com.github.correa.finalreality.enums.Stats;
 import com.github.correa.finalreality.model.character.ICharacter;
 import com.github.correa.finalreality.model.character.player.classes.AbstractMage;
 import com.github.correa.finalreality.model.weapon.IWeapon;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.HashMap;
 import java.util.Objects;
 import java.util.concurrent.BlockingQueue;
 
@@ -17,17 +20,19 @@ public class WhiteMage extends AbstractMage {
    *  the queue with the characters waiting for their turn.
    * @param name
    *  the White Mage's name.
-   * @param hitPoints
-   *  the White Mage's hit points.
+   * @param maxHitPoints
+   *  the White Mage's maximum hit points.
    * @param defensePoints
    *  the White Mage's defense points.
+   * @param maxManaPoints
+   *  the mage's maximum mana points.
    */
   public WhiteMage(@NotNull BlockingQueue<ICharacter> turnsQueue,
-      @NotNull String name, int hitPoints,
-      int defensePoints, int manaPoints) {
+      @NotNull String name, int maxHitPoints,
+      int defensePoints, int maxManaPoints) {
     super(
-        turnsQueue, name, hitPoints,
-        defensePoints, manaPoints);
+        turnsQueue, name, maxHitPoints,
+        defensePoints, maxManaPoints);
   }
 
   @Override
@@ -38,6 +43,15 @@ public class WhiteMage extends AbstractMage {
   @Override
   public void equip(final @NotNull IWeapon weapon) {
     weapon.equippedByWhiteMage(this);
+  }
+
+  @Override
+  public HashMap<Stats, String> getInfo() {
+    super.getInfo();
+    info.put(
+        Stats.CHARACTER_TYPE,
+        String.valueOf(CharacterType.WHITE_MAGE));
+    return info;
   }
 
   @Override