+++ src/main/java/com/github/correa/finalreality/model/weapon/commonweapons/Axe.java
@@ -1,9 +1,12 @@
 package com.github.correa.finalreality.model.weapon.commonweapons;
 
+import com.github.correa.finalreality.enums.Stats;
+import com.github.correa.finalreality.enums.WeaponType;
 import com.github.correa.finalreality.model.character.player.classes.commonclasses.Engineer;
 import com.github.correa.finalreality.model.character.player.classes.commonclasses.Knight;
 import com.github.correa.finalreality.model.weapon.AbstractWeapon;
 
+import java.util.HashMap;
 import java.util.Objects;
 
 public class Axe extends AbstractWeapon {
@@ -29,6 +32,15 @@ public class Axe extends AbstractWeapon {
   @Override
   public void equippedByKnight(Knight knight) {
     knight.equippedBy(this);
+  }
+
+  @Override
+  public HashMap<Stats, String> getInfo() {
+    super.getInfo();
+    info.put(
+        Stats.WEAPON_TYPE,
+        String.valueOf(WeaponType.AXE));
+    return info;
   }
 
   @Override