+++ src/test/java/com/github/cc3002/finalreality/model/weapon/types/AxeTest.java
@@ -1,9 +1,14 @@
 package com.github.cc3002.finalreality.model.weapon.types;
 
 import com.github.cc3002.finalreality.model.weapon.AbstractWeaponTest;
+import com.github.correa.finalreality.enums.Stats;
+import com.github.correa.finalreality.enums.WeaponType;
+import com.github.correa.finalreality.model.weapon.IWeapon;
 import com.github.correa.finalreality.model.weapon.commonweapons.Axe;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * Set of tests for the Axe class.
@@ -24,6 +29,7 @@ public class AxeTest extends AbstractWeaponTest {
   @BeforeEach
   void setUp() {
     expectedAxe = new Axe(AXE_NAME, DAMAGE, SPEED);
+    testWeapon = expectedAxe;
   }
 
   @Override
@@ -32,6 +38,22 @@ public class AxeTest extends AbstractWeaponTest {
     var testEqualAxe = expectedAxe;
     checkIWeaponConstruction( expectedAxe,
         testEqualAxe, new Axe("Test", DAMAGE, SPEED));
+  }
+
+  @Override
+  @Test
+  protected void getInfoTest() {
+    checkGetInfo(testWeapon);
+  }
+
+  @Override
+  protected void checkGetInfo(IWeapon weapon) {
+    super.checkGetInfo(weapon);
+    var testInfo = weapon.getInfo();
+    assertEquals(
+        WeaponType.AXE,
+        WeaponType.valueOf(
+            testInfo.get(Stats.WEAPON_TYPE)));
   }
 
 }