+++ src/test/java/com/github/cc3002/finalreality/model/weapon/types/SwordTest.java
@@ -1,9 +1,14 @@
 package com.github.cc3002.finalreality.model.weapon.types;
 
 import com.github.cc3002.finalreality.model.weapon.AbstractWeaponTest;
+import com.github.correa.finalreality.enums.Stats;
+import com.github.correa.finalreality.enums.WeaponType;
+import com.github.correa.finalreality.model.weapon.IWeapon;
 import com.github.correa.finalreality.model.weapon.commonweapons.Sword;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
  * Set of tests for the Sword class.
@@ -24,6 +29,7 @@ public class SwordTest extends AbstractWeaponTest {
   @BeforeEach
   void setUp() {
     expectedSword = new Sword(SWORD_NAME, DAMAGE, SPEED);
+    testWeapon = expectedSword;
   }
 
   @Override
@@ -32,5 +38,21 @@ public class SwordTest extends AbstractWeaponTest {
     var testEqualSword = expectedSword;
     checkIWeaponConstruction( expectedSword,
         testEqualSword, new Sword("Test", DAMAGE, SPEED));
+  }
+
+  @Override
+  @Test
+  protected void getInfoTest() {
+    checkGetInfo(testWeapon);
+  }
+
+  @Override
+  protected void checkGetInfo(IWeapon weapon) {
+    super.checkGetInfo(weapon);
+    var testInfo = weapon.getInfo();
+    assertEquals(
+        WeaponType.SWORD,
+        WeaponType.valueOf(
+            testInfo.get(Stats.WEAPON_TYPE)));
   }
 }